Modify update-all-windows to also apply to windows in other frames.
Instead of starting with the result of calling window-list, which is
only applicable to the current frame, we instead start with
buffer-list and then filter that based on whether it's associated with
a visible window, regardless of which frame it happens to reside in.
Once we have that list, we can apply the same check for
git-gutter-mode before invoking git-gutter on the buffer.